Update contact.html
Add an image to represent "communication" on the right to make it seem like column 2. In this case a phone icon best represents coming into contact with someone. Images bring attention so the user will look at the image first before reading the form. Allows users to be directed correctly.