Update contact.html
Previous change and current change is the Q and A section that was originally used in the first turn in. This section allows users to see other questions that were asked by users to see if their own questions have been answered. If their question has been answered, they will not need to submit the form.